Segments: store size in 32 bit int
fixes
https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=38876&sort=-opened&can=1&q=proj%3Agrok

+++ src/lib/jp2/t1/Codeblock.h
@@ -206,7 +206,6 @@ struct DecompressCodeblock : public Codeblock
 			for(uint32_t i = 0; i < numSegmentsAllocated; ++i)
 				new_segs[i] = segs[i];
 			numSegmentsAllocated *= 2;
-			;
 			delete[] segs;
 			segs = new_segs;
 		}
@@ -261,8 +260,8 @@ struct DecompressCodeblock : public Codeblock
 
   private:
 	Segment* segs; /* information on segments */
-	uint16_t numSegments; /* number of segment in block*/
-	uint16_t numSegmentsAllocated; // number of segments allocated for segs array
+	uint32_t numSegments; /* number of segment in block*/
+	uint32_t numSegmentsAllocated; // number of segments allocated for segs array
 };
 
 } // namespace grk
